Move write_text from ExpandedFst to SerializableFst
