+++ rustfst/src/fst_impls/const_fst/misc.rs
@@ -2,6 +2,6 @@ use std::fmt;
 
 use crate::fst_impls::ConstFst;
 use crate::fst_traits::{ArcIterator, CoreFst, FinalStatesIterator, StateIterator};
-use crate::semirings::Semiring;
+use crate::semirings::SerializableSemiring;
 
 display_fst_trait!(W, ConstFst<W>);
