+++ rustfst/src/fst_impls/vector_fst/misc.rs
@@ -2,6 +2,6 @@ use std::fmt;
 
 use crate::fst_impls::VectorFst;
 use crate::fst_traits::{ArcIterator, CoreFst, FinalStatesIterator, StateIterator};
-use crate::semirings::Semiring;
+use crate::semirings::SerializableSemiring;
 
 display_fst_trait!(W, VectorFst<W>);
