+++ rustfst/src/semirings/macros.rs
@@ -1,0 +1,42 @@
+macro_rules! test_semiring_serializable {
+    ($sem_name:  tt, $( $weight: expr )* ) => {
+
+        #[cfg(test)]
+        mod $sem_name {
+            use super::*;
+
+            #[test]
+            fn test_serializable_binary() -> Fallible<()> {
+                for weight in &[ $( $weight ),* ] {
+                    let weight = weight.clone();
+                    let mut serialization = vec![];
+
+                    weight.write_binary(&mut serialization)?;
+
+                    let (_, weight_deserialized) = SerializableSemiring::parse_binary(serialization.as_slice())
+                        .map_err(|e| format_err!("Can't parse weight : {:?}", e))?;
+
+                    assert_eq!(weight, weight_deserialized);
+                }
+
+                Ok(())
+            }
+
+            #[test]
+            fn test_serializable_text() -> Fallible<()> {
+                for weight in &[ $( $weight ),* ] {
+                    let weight = weight.clone();
+                    let serialization = format!("{}", weight);
+
+                    let (_, weight_deserialized) = SerializableSemiring::parse_text(serialization.as_str())
+                        .map_err(|e| format_err!("Can't parse weight : {:?}", e))?;
+
+                    assert_eq!(weight, weight_deserialized);
+                }
+
+                Ok(())
+            }
+        }
+
+    };
+}
