+++ rustfst/src/semirings/mod.rs
@@ -1,5 +1,9 @@
 #[macro_use]
 mod semiring;
+
+#[macro_use]
+mod macros;
+
 mod boolean_weight;
 mod gallic_weight;
 mod integer_weight;
