+++ rustfst/src/tests_openfst/algorithms/reverse.rs
@@ -7,7 +7,7 @@ use crate::algorithms::MapFinalAction;
 use crate::algorithms::WeightConverter;
 use crate::algorithms::{reverse, weight_convert};
 use crate::fst_impls::VectorFst;
-use crate::fst_traits::{AllocableFst, MutableFst, SerializableFst};
+use crate::fst_traits::{AllocableFst, CoreFst, MutableFst, SerializableFst};
 use crate::semirings::WeaklyDivisibleSemiring;
 use crate::semirings::{Semiring, SerializableSemiring};
 use crate::Arc;
@@ -47,6 +47,7 @@ pub fn test_reverse<F>(test_data: &FstTestData<F>) -> Fallible<()>
 where
     F: SerializableFst + MutableFst + AllocableFst + Display,
     F::W: 'static + SerializableSemiring + WeaklyDivisibleSemiring,
+    <<F as CoreFst>::W as Semiring>::ReverseWeight: SerializableSemiring,
 {
     let fst_reverse: VectorFst<_> = reverse(&test_data.raw).unwrap();
     let mut mapper = ReverseWeightConverter {};
